Fix bugs with zero?/positive? money
Previously we changed uses of '== 0' and '> 0' to 'zero?' and
'positive?', but this has lead to a bug since the object receiving the
method calls is actually an instance of SmartAnswers::Money, which
doesn't support them. Rather than risk changing the behaviour of this
class by supporting these methods, this reverts the change and disables
the NumericPredicate cop where instances of this class are in use.