Update logic for compressed-hours calculations
Compressed hours is calculated the same as hours worked per week calculations.
Therefore, there was no need to add new logic but we have updated the method name to
reflect this and updated the test description.

Co-Authors: @deborahchua @callumknights

Add method to show hours/minutes and not decimals

The method `full_time_part_time_hours_and_minutes` is used in the flow
and is needed to show the entitlement time in hours and minutes instead
of decimals: `[2, 30]` vs `2.5`.