Add rspec-rails gem + config & "default" rake task
This is the basis to start migrating from minitest to rspec.

The `rspec-rails` gem has been added, plus the related configuration in
the `spec_helper.rb` file.

In the Rakefile, the "default" rake task has been overridden so that it
now runs both minitest and rspec tests. This allow us to use:
- `rake test` to only run minitest tests
- `rake spec` or `rspec` to only run rspec tests
- `rake` to run both minitest and rspec tests