Update shifts flow
- Update wording in the outcome
- Add a new partial to accommodate the new wording about the employer
- Update question numbers to reflect what we have in the [Calculator Logic Tree](https://docs.google.com/presentation/d/17NaE8677fpdkq67AUwDTZ6fTXy0sy3mi8T6T80DC7I8/edit#slide=id.p13)
- Remove the use of `hours_per_shift` in the class as it's only used
as part of the outcome wording at the end
- Update test and use latest test data

Update shifts "starting" flow

- Update wording in the outcome to render additional partial for
this scenario only
- Update question numbers to reflect what we have in the [Calculator Logic Tree](https://docs.google.com/presentation/d/17NaE8677fpdkq67AUwDTZ6fTXy0sy3mi8T6T80DC7I8/edit#slide=id.p13)
- Update test and with lastest test data

Co-authored-by: Ed Kerry <edward.kerry@digital.cabinet-office.gov.uk>

Update shifts "leaving" flow

- Switch to using the `shift_entitlement` which calls the
`rounded_shift_entitlement`, updated in a previous PR
- Update tests and with latest test data
- Update question numbers to reflect what we have in the
Calculator Logic Tree

Co-authored-by: Ed Kerry <edward.kerry@digital.cabinet-office.gov.uk>

Update shifts "starting-and-leaving" flow

- Update test and with latest test data

Add additional message for shifts flow

This message is displayed if the number of shifts worked per week
is over 5 shifts.

Fix shift additional message

In order for this message to be displayed it requires one of the
methods which is currently private.