Update maternity calculator's rate changes dates
Rate changes happen on the first Sunday of April each year. This doesn't
necessarily coincide with the start of the fiscal or financial year.

For the moment, this only updates
`lib/data/rates/maternity_paternity_adoption.yml`
and not
`lib/data/rates/maternity_paternity_birth.yml`
because we're waiting for an update from HMRC on whether the dates there
should be different and why some of them are in July instead of April.

Trello card: https://trello.com/c/9v5Az2SA/1697-3-employers-maternity-calculator-built-on-tax-years