Script to update marriage abroad outcomes
This is a one-off script is to bulk update `_opposite_sex.govspeak.erb`
files for each country.

The content team have proposed 7 variations of text to add:
1 - commission text with link
2 - embassy text with link
3 - government text
4 - commission text with link (above ##Prove you're free to get married)
5 - embassy text with link (above ##Prove you're free to get married)
6 - embassy text with link (above ##What you need)
7 - consulate text with link (above ##What you need to do)

They are provide in the csv file, which the script uses to determine
how to update the various files for each country. An example of a csv entry:
`Canada,1,/world/organisations/british-high-commission-ottawa,`

The script only apply changes to one country at a time (the first row it
finds where `updated` is not set to `true`, so that it's easier to review the
changes and separate them into individual commits. If the changes have been
successfully applied the `updated` field is set to `true` and will be
ignored in the next run of the script.

Run with `rails r script/update-marriage-abroad-opposite-sex-outcomes.rb`