Fix render file deprecation warning in layouts
We were receiving this deprecation warning:
`DEPRECATION WARNING: render file: should be given the absolute path to 
a file`

It seems that because the file we are rendering is in application we can
use the `render(template: )` method instead.

`render(file: )` does not reply on rails look up logic, and therefore
needs a absolute path. Here we can rely more on the rails look up logic
by is using `render(template: )`` since the layouts are held in the
app/views folder.

Co-Authored-By: Rebecca Pearce <rebecca.pearce@digital.cabinet-office.gov.uk>