Fix minimum wage calculator rounding issues
It was discovered as part of the uprating work that the minimum wage calculator
could give false positive results due to it needing to round the minimum wage of
a given user to 2 decimal places to finish it's calculation. This meant in some
cases the total pay a person recieved in a given time period meant they were
earning slightly under mininum wage (maximum 0.5 pence per hour under).

To rectify this I changed the calculator to use the minimum entitlement for
a given number of hours as it's basis for deciding whether a user is being
above or below the minimum wage. This is more suitable as the total pay a user
has recieved is given to 2 decimal places in the first place, and the minimum
wage in the `rates.yml`, which minimum entitlement is based off is also given
to 2 decimal places. This avoids any unneccessary rounding and makes the
calculator more accurate in cases where the employee is earning an amount very
close to minimum wage.

I have also added some test cases to the flow testing, just to cover some extra
scenarios using accurate numbers given to me by content.

Trello card > https://trello.com/c/k8xvMhKs/1813-minimum-wage-calculator-bug