Fix Smart Answers Rails 6 issue
Noticed when releasing some separate smart-answers that a bug had been
introduced by the rails 6 upgrade, this was a production/staging only
bug so likely not caught.

The issue was that with the introduction of Zeitwork rails 6 expects
Structs to be defined inside a `Data` module, and where we weren't doing
this Zeitwork was throwing an error.

The easiest solution to this seemed to be to move the offending files out
of the `/data` folder and into the main `/lib` path.