Remove call out
Remove %The service to register an overseas death with the UK authorities is currently suspended. This page will be updated when it has resumed.% 
Because it draws the user's eye down to that warning. Users may not realise they can use the tool to find out how to register a death in the UK. The majority of users will be registering a death in the UK.
The General Register Office asked for it to be removed. FCO - who had asked for it to be added - agreed that it could be removed