Add initial flow for coronavirus business support
This adds a draft smart answer flow for supporting businesses during
the Coronavirus pandemic. Content is placeholder only, and we intend
to iterate this and add tests once the initial code is merged. Using
a 'draft' status means this will only be visible in development [1].

Since the purpose of this smart answer is to filter from a number of
potential results, we've chosen to use a single 'results' outcome page,
and a Calculator class [2] to determine which of the results to display
to the user; each result has its own bespoke criteria to display.

[1]: https://github.com/alphagov/smart-answers/blob/master/doc/smart-answers/flow-definition.md#metadata
[2]: https://github.com/alphagov/smart-answers/blob/master/doc/smart-answer-flow-development/refactoring.md