Extend example of how to write questions
Previously I struggled to write my first smart answer because I didn't
appreciate the need to specify options in two places (in the Flow, and
an optionally in the view, to change the labels). This increases the
signposting to the example code in the doc, and extends the example to
show how the view and the Flow class work together.