Remove draft options from FlowRegistry
These draft options prevented the viewing of smart answers when content
wasn't flagged as published. These constraints aren't necessary as the
draft handling can be done by whether this content is published in the
Publishing API or not.

Doing this allows removing the FLOW_REGISTRY_OPTIONS constant that is
modified which was somewhat unconventional. This has instead been
replaced with the ability to set FlowRegistry options when replacing the
instance.

I expect there are means that could be applied to simplify this further
as it feels rather complex but I'm not too familiar with the code base
at this point.