Allow draft flows to be presented to draft GOV.UK
This renames the publish method on ContentItemPublisher to sync and
adapts the Publishing API calls so that any flows that are defined with
a status of draft are sent to the Publishing API but only published
items are published. This is to enable the draft previewing of smart
answers.

I chose to rename the publish method to sync to reflect the wider
functionality this has.