Rename ContentItemPublisher to ContentItemSyncer
The naming of ContentItemPublisher no longer seemed appropriate since this
class only had a lone method of sync. I'm not super happy with the new name
of this class but I couldn't think of a better name and I think my issues
lie more in it's usage and purpose than name that could maybe be refactored
away.