Remove outdated/inconsistent development documents
The information in these docs is now out of date and in danger of
misleading developers. The situations that are documented are common
across most GOV.UK apps so they don't seem to deserve dedicated
resources in this repo.

This is a bit of a shame as obviously these were nicely written docs but
I think they're likely to slip out of date again.