Remove heroku startup script
This has now been superseded with the review app system provided by
Heroku. I'm removing this to simplify the application so there are less
choices for developers when working with this and as part of work
cleaning up legacy cruft built up in this repo.

I've also updated the fact-check documentation to remove references to
this script and remove out-of-date information.