Move smart answer configuration out of lib/data
Since this is configuration information it feels somewhat out of place
within the lib directory of this project where you would anticipate
finding code.

I chose to rename the directory from data to smart_answers so that this
information is more clearly config for smart answers when looking at the
directory. The name "data" felt rather ubiquitous. The naming was a pain
though as the only clear convention is that these files are yaml.

Moving this resolves a deprecation warning where Zeitwerk tries to load
a Ruby constant of Data. This could probably have been resolved by
removing this path from being eager loaded.