Re-arrange integration test for clarity of purpose
This adjusts the context of this test to be more focused on what we're
actually testing. In this case we've got two scenarios that lead to
either a results or no results page.

I've made the code out of setup calls as it didn't feel idiomatic for us
to be making assertions during setup.