Replace self_employed question with paye_scheme
This adds a new question taking the slot of question 4 to check whether
an employer is part of PAYE schema.

This change had the side effect of setting two pieces of guidance
that'll appear in all scenarios so I've removed the no results code. I
know at least one of these will be removed so I'm hoping that I don't
need to re-add this code in the next couple of hours but hey better than
it accidentally sitting there dormant.