Remove outcome_flattener config
This seems to be a relic from when there were far more outcome
flatteners. As there is now only the ability to flatten marriage-abroad
content this doesn't seem necessary any more.

I couldn't seem to work out why it needed the time set to a particular
date so I felt it was fine to remove the timecop call - this is pretty
wacky in application code anyway. It also seemed rather confusing having
a method called current time that returned a string of "2016-10-01" with
no rhyme or reason as to why that date special.

The reason for this choice of date in the first place seems to be
explained in https://github.com/alphagov/smart-answers/commit/eb89ea447ee84fdb23062680c9d364d383edbdf8
this seems to indicate that having a very old data was a problem. Given
these files are no longer used for regression testing I don't think this
being the current time is a particular concern.