Use a capitalised none when showing answered questions
As the more prevalent (but unfortunately not universal) convention is for
questions to have capitalised answers, it can be a bit jarring when you
select the none option on a checkbox question and have the previous
answer shown without capitalisation.

This amends this by hardcoding the response as a capitalised string.

I figured there wasn't much value in keeping this tied around the value
variable provided, since it can't vary, and was clearer as an explicit
string.