Remove fake content store
We can avoid the need for this by using the real content store with
destination set by environment variable. This environment variable and
the one that existed before are moved to be set conditionally in the
task. This was done because it seemed rather strange to be setting env
vars (and other stubs) in an application class.