Remove lib/services.rb
This file is rather unnecessary as it was already using short methods to
look up API adapters and just storing them to class variables (which
is somewhat dubious for thread safety).

It's also one of my least favourite patterns of coding in GOV.UK as this
app (like most) also has a concept of services which is very different
to these.