Move GdsApi test helpers to test_helper class
There are a large number of repetitive require and includes throughout
the test suite and I'm risking the need to add to add a lot more by
wanting to use GdsApi test helpers to stub content store requests.

To cut down on all this duplication I've moved all of these to be
defined in the test_helper.rb file.