Light refactor of world location tests
This makes a few basic amends to the world location tests with the main
goal being to replace object stubs with API stubs.

It seems we were doing some unnecessary API assertions and weren't
making use of storing our created stubs to test against. There were also
some places that should be asserting the API was called (such as testing
loading all world locations)

I also used this as an opportunity to simplify the time usage in
Timecop.