Adds minor fix to date parsing and future logic
Minor fix in that we weren't considering the 16th as a valid date when
discussing shielding pay.

Future logic for when the user has shielding but returns to work before
the 16th of April. This should not be possible but currently we have no
outcome for this, so this commit includes underlying logic for when
content is made aware.