Manually fix remaining linting errors
This exposed a conflict between two Cops, presumably due to the
esoteric code in this file, which the authors weren't expecting.

- Style/EmptyMethod
- Style/Semicolon

Trying to fix Style/Semicolon by putting the 'end' on a newline
leads to an issue with Style/EmptyMethod. Disabling these specific
Cops also doesn't work for some reason: RuboCop says it's "unnecessary".
This disables all RuboCop checks are the offending empty methods.