Remove html argument in content_for
This parameter provided the option to retrieve content without
processing it as govspeak i.e `html: false` prevents the block being
processed as govspeak and returns plain text. This is no longer needed
as it was used for a debugging feature to user unprocessed govspeak,
which was removed in earlier commits.