Add FormatCaptureHelper to render content_for blocks
The FormatCaptureHelper was added to render content_for blocks within
smart answer templates. Previously it was not possible to specify the
content type for particular sections within the template, as they were
all post-processed in the same way. e.g. content for the body was
always passed through the Govspeak processor before being rendered in
the main views (result, question, start). Therefore there was no way to
specify a content format within the node template. e.g. html content for
the body, not processed by govspeak. This helper enables the option to
specify a content format with the content_for function within the node
template.

To enable this we are also expanding the view paths of ErbRenderer to
include those from Rail's ActionController::Base. This is because we
need access GovukPublishingComponents to fully render govspeak.