Refactor FormatCaptureHelper#content_for to render_content_for
Previously this method was intented to override the `content_for`
within ActiveView::Helpers:CaptureHelper. This changed the expected
behaviour of content_for within template and could lead to confusion. It
also put dependency on undocumented behaviour of the content_for within
Rails, which could be fragile and confusing to support in the future.

This change requires an update to all templates to use the renamed method.