Fix error for empty render_content_for blocks
This fixes the case when a template has render_content_for block that is
empty or just has whitespace. This was seen for for UK Visa smart answer
where the body content is empty for EEA citizens. `render_content_for`
expected the block to always be present and return an String. However
the erb templating engine passes a block that evaluates to a nil value
instead of a empty string.