Add declarative smart answer helper methods
This adds `text_for`, `govspeak_for` and `html_for` methods intended
for the smart answer flow methods as a way for users to define content
for a smart answer. These are intended to replace the
`render_content_for` helper that has logic to determine the content
type.

This is set-up to be backwards compatible with the render_content_for
method so that both can co-exist for a period of time to allow
migration.

In doing this I also simplified a few things. We don't make any use of
options when calling `content_for` so these methods don't need to
support an option. All our usages of these methods use a block to add
content so I've set it up to raise an error.

I also removed the need to remove excess new lines after govspeak is
rendered. Since we already have HTML at that point I'm not sure what
value that adds.