Update documentation for the different helper methods
This updates the documentation to remove references to .govspeak.erb
files and to make suggestions of which helper method (text_for,
govspeak_for, html_for) in each scenario.

Along the way I noticed that the label field may not be working (issue
prior to these changes) so I left a note about that so the next person
who wants it can investigate.

I removed the information on error_* as I believe that is mostly set on
a custom basis so doesn't need to be included in the general
documentation.