Raise errors when setting HTML for text content
This adds in a check when specifying HTML in a smart answer (via
govspeak_for or html_for) that it is not for a field that is text only.
This is put in to address a concern that it could be rather easy for
someone to specify `govspeak_for(:title) { "my govspeak" }` and this
produce an unexpected output that no-one notices.

As part of this I removed the govspeak option in DEFAULT_FORMATS as I
didn't think it affected the outcome of any method calls.