Switch asset path and host
This configures this app to serve assets from the current host by
default rather than the GOVUK_ASSET_ROOT environment variable. This is
to support the switch of assets from assets.publishing.service.gov.uk to
www.gov.uk.

The host for assets can be specified with an ASSET_HOST env var which is
intended to be used in development environments where router isn't set
up for proxying. Unlike GOVUK_ASSET_ROOT this is expected to be app
specific hence why it doesn't have a GOVUK prefix.