Add internal link tracking helper
This adds a JS function that sets up GA tracking for internal links i.e.
links within the same domain. It add an event listener to each link with
a relevant href and sends a GA event if clicked. This is called on results
pages to track user clicks and is needed by certain smart answer flows e.g.
business-coronavirus-support-finder. This helper function makes it
easier to enable internal link tracker, instead of having to manually
add the correct data attribute to links to enable tracking. It will also
help prevent future links from being missing in tracking, if the data
attribute are forgotten to be added.