Remove last updated date on pages
This removes Last updated date on the bottom of smart answer pages as it
is inaccurate and potentially confusing for user. This feature was
implemented to show the last updated date to be the date when the
application was last deployed. This means the last_updated date
refelected any change to any smart answer and may have not even been a
user facing change. This may give the user a false sense of confidence
that the content in the smart answer is upto date.