Replace 'select' input with Design System's 'date_input'
The Service Manual states that a [text input pattern is best][manual]
for DOB and other date inputs. This came up in benchmarking the
State Pension smart answer; as years are rendered chronologically
inside the `<select>`, the first year to show as an option is
1895, and it's fiddly for users to scroll to their actual birth year.

This commit does the bare minimum to swap out the `<select>`
inputs with the `date_input` component provided by the Design
System:
https://design-system.service.gov.uk/components/date-input/

Some of the logic can be moved out of the view, but we will do
that in later commits.

I did try only passing `{ name: "response" }` to the component,
which would have worked for most of the replacement, but Smart
Answers sometimes doesn't render the day, month or year if a
default one is provided, so we needed to be able to conditionally
render each input. I've therefore had to pass an `items` array.

Smart Answers also - sometimes - wants to prefill the values for
day/month/year, so I've done a `reduce` on the `X_options` provided
by the presenter to retrieve this value. If it can't be found,
`nil` is returned and no value is filled on the input.

https://trello.com/c/6erUcSo8/2020-3-investigate-changing-format-of-date-questions-in-smart-answers

[manual]: https://www.gov.uk/service-manual/design/form-structure#structure-your-form-to-help-users