Simplify data_question_presenter.rb
We no longer need the days_options, months_options and years_options
methods, which were used for populating the `<option>`s inside the
`<select>`s that we've now replaced with the [date-input] component.

We were only using it for determining whether or not the Day, Month
or Year input should be pre-filled with a value. This has now been
stripped down to its bare parts.

We've lost some logic in terms of the default range for the 'year'
if not provided. This was from "1 year ago" to "3 years from now".
There are no tests reinforcing this logic, and there's nothing
stopping the user from entering a year inside or outside that range.
I assume it was only there to provide _something_ inside the year's
`<select>` field in case nothing was specific. Given that we're
now using a text input, this isn't something we have to worry about
anymore.

[date-input]: https://components.publishing.service.gov.uk/component-guide/date_input