Use content ids to associate Smart Answers with needs
This follows on from some work done a long time ago to move needs from
the Need API to the Publishing API [1]. At this time, the identifiers
for needs transitioned from being numbers, to UUIDs, because the
Publishing API uses UUIDs.

1: https://github.com/alphagov/maslow/pull/160

I don't think this change alone will actually have an effect beyond
smart answers, but it may enable linking the smart answers content
items to the relevant need content item in the future.

To map the need ids here to content ids, I just queried the Publishing
API database with a query like the following, where XXX is the numeric
need id:

  SELECT DISTINCT editions.details->'need_id', documents.content_id
  FROM editions
  INNER JOIN documents
    ON editions.document_id = documents.id
  WHERE document_type = 'need'
    AND editions.details->>'need_id' = 'XXX';