Add Question 1 to logic flow
Add question to enter the number of children the claim is being made
for. This uses the `value_question` question type and we have set
it to only parse integer values. For more information see https://github.com/alphagov/smart-answers/blob/master/doc/smart-answers/question-types.md#value_question

In Calculators it is a drop down with the maximum number the user
can select is `10`. By changing to an input field we can increase
the limit in the backend code.

For accessibility reasons using a select drop-down
is not recommended - see https://design-system.service.gov.uk/components/select/#when-to-use-this-component

Co-authored-by: Deborah Chua <deborah.chua@digital.cabinet-office.gov.uk>