Add validation to the flow inputs
Add methods to the calculator to validate the inputs from the flow
questions as well as add in the custom validation messages in the
question views.

Back in 2012 we originally built this calculator in smart answers
and the limit set for the number of children the user can enter
was set to `30` (see https://github.com/alphagov/smart-answers/pull/513/files#diff-27b83c907831b5a7ca4d43b931e3ec41L136)
We decided to use this limit for `Q1` but this can be changed later
if this value is not suitable.

Co-authored-by: Deborah Chua <deborah.chua@digital.cabinet-office.gov.uk>