Fix escaped text being sent to Publishing API
This marks content within a text_for block in a ERB smart answer
template as HTML safe (e.g. title, meta description). This prevents
characters being escaped, which can be sent to Publishing API and cause
the escaped characters to be displayed. For example the use of
apostrophes in a title.