Add FCDO to sponsoring organisation check
This is part of the work around the Machinery of Government changes that
will see FCO and DFID merge to become a new department.

After noticing that [a run in integration](https://deploy.integration.publishing.service.gov.uk/job/run-rake-task/172527/console)
of the rake task introduced in
https://github.com/alphagov/whitehall/pull/5777 had a breaking effect on
some smart answers, we decided to check for both FCO and FCDO in
the `WorldwideOrganisation#fco_sponsored?` method.

After all content has been switched to the new FCDO organisation there
will be some cleanup work to do (e.g. remove FCO, rename methods, etc),
but for the moment this allows us to go ahead with the sponsoring
organisations work.

Trello card: https://trello.com/c/yBvz6SbU/2095-spike-updating-worldwide-organisation-sponsoring-organisations-automatically-to-a-new-org-replace-fco-with-fcdo