Merge FlowRegistrationPresenter into FlowPresenter
The functionality from FlowRegistrationPresenter has been merged with
the FlowPresenter. The FlowPresenter is now used to construct Content
Item payloads. This helps fix the issue where duplication of methods was
required in both presenters to generate attribute for the views and the
content item. This made it possible for difference between the content
item and the rendered view. For example the landing page had a different
start link in content item vs the rendered page by smart answers.