Fix no-results text & sub-section display issues
This change fixes the following two issues that are affecting users of
the Coronavirus Find Support flow result page.

1. It is possible to reach the results page without the no-results text
being shown.

An example of this behaviour is:

  * "What do you need help with because of coronavirus?" : "Feeling
    unsafe where you live, or what to do if you’re worried about the
    safety of another adult or child"
  * "Do you feel safe where you live?" : "Yes"
  * "Where do you want to find information about?" : "England"

The expected response is that the user is shown the no-results text.
However, a blank page is shown.

2. Sub-sections are not being correctly shown/hidden as appropriate
and based on the users given responses.

An example of this behaviour is:

  * "What do you need help with because of coronavirus?" : "Getting
    food"
  * "Are you finding it hard to afford food?" : "Yes"
  * "Are you able to get food?" : "Yes"
  * "Where do you want to find information about?" : "England"

The expected response is that the user is shown the "Getting food"
group and the "If you’re finding it hard to afford food" sub-section,
but not the "If you’re unable to get food" sub-section.  However, both
sub-sections are displayed.