Fix change your answer links in session flows
Replacing the param :node_name with :node_slug meant node name was not
being passed to flow presenter. This fix passes the node name into params
before they are passed to the flow presenter.

`params[:node_name]` is passed from `FlowPresenter` to `Flow` where it is
used in the method `resolve_state` as `requested_node`. It needs to be
present for the correct current state to be identified

An integration test is also added that ensures that the "Change" behaviour
works as expected.