Update Cache-Control headers for session answers requests
This is consistent with the header on govuk-coronovirus-find-support
which we know work.

We need to add max-age=0 because fastly (our CDN) doesn't obey
"no-store, no-cache".

Private also prevents cache of any intermediate servers.

must-revalidate tells the browser and servers to never use stale content
(potentially when errors are served)

Pragma is only for backwards compatibility with HTTP/1.0 clients - which
no longer applies as we don't support HTTP/1.0 connections.

Expires can be replaced with max-age=0