Fix the "Not sure" response to need_help_with flow
What

The "Not sure" option on the "What do you need help with?" is meant to
route the user through to all question groups. However at the moment it
skips all the questions and goes straight to the "Nations" question.

Why

To make the behaviour consistent with the current tool.