Remove previous answers from results page
What

We should remove the previous answers section on the results page, but
keep the start again link. This makes the behaviour of navigation
consistent with the current tool.

Why

There is a concern about the behaviour of the previous answers section on
the results page of the find coronavirus support flow. This flow uses
sessions, which means answers to all questions are remembered and the user
does not have to re-answer subsequent questions if going back in the flow
to change a answer. This could be potentially confusing if the user
expects to see related questions that are dependent on an answer to an
earlier question, i.e. need help with -> getting food, expects to see
questions about getting food.