Use GovukTest to provide headless chrome selenium options
This changes the configuration mechanism for Jasmine selenium runner to
use the approach advised in the gem (as per:
https://github.com/alphagov/content-publisher/pull/2110) and also uses
GovukTest as a means to share the headless chrome configuration.

This change is made to provide govuk-docker compatibility in upcoming
changes: https://github.com/alphagov/govuk-docker/pull/394

It also removes some unnecessary version numbers from the Gemfile.