Move 'Start again' button on results page
What

The current position of the 'Start again' link on the results page of
the find support smart answer breaks up the flow of the text.

We should move it to the bottom of the page, below a header for
'Change your answers'.

Why

So that users can find what they need to know first, before selecting
'Start again'.

So that people know why they would start again.