Make class methods in BenefitCapCalculatorConfiguration
This move the existing instance methods to class methods. Additionally,
the exempt_benefits attribute is now a function parameter. This is to
allow the class to be saved to the state object, for use within
templates. This will make it easily to refactor the flow definition and
remove precalculate methods.