Redirect find-support users who don't need help to results page
This changes the previous behvaiour where if user's selected "None of
these/not sure" option on the "What do you need help with?" question
they'd be shown every subsequent question. Now the user skips the
question about which particular section they need help with and taken
directly to the "Where do you live?" question and the no results page.