Add "or" divider above none options in checkbox questions
This adds the word "or" to seperate the exclusive none option from the
rest. This helps give users who use a screen reader more context about
the behaviour of selecting the none option. This is becuase they might
not be notified by the screen reader that previous option were
automatically deselected when selecting "none".