Match EU experience with non-visa nationals
Where EU citizens are arriving from 2021, the experience should
match that of non-visa nationals. I've added some logic to the
title to ensure that the "You do not currently need a visa" title
is displayed when it should be, rather than the EU default "You do
not need a visa to come to the UK".

There are no tests around any of the conditionals in this template.
In the next commits I'll start splitting this template out into
logicless templates, where we can explicitly test against the
`outcome` and have more confidence that everything works correctly.