Create new 'no tourist visa needed' outcome
The text for this outcome matches that of outcome_school_n.erb,
but without the reference to schools. It seems to make more sense
to split these out rather than attempt to DRY up the outcomes and
use conditionals within one template.

Non visa nationals used to see the 'no visa needed for school'
outcome, as it didn't have any school specific content at the time:
https://github.com/alphagov/smart-answers/commit/ba43f4f0c97085d012ef8585b790d0b990eeff41#diff-099c4e1288e8c7007f0947bf1701df3e4f6e9efc7b34f927533210d001b49f78R255

However, in September the content was changed to include some
school content:
https://github.com/alphagov/smart-answers/commit/1991a091ff69d1c3783a2e163556d0dac8302c15

We therefore needed to split out the outcomes logically.