Always show sector question in business support finder
Previously the sector question was shown to users who'd not answered
"none" to the "What is the rateable value of your business' non-domestic
property?" question. We are flattening the flow to accomodate the
sector question to be used by other schemes where there is no explicit
eligibity around having non-domestic rateable property.