Modernise Rakefile and increase consistency with other GOV.UK apps
This removes the executable flag and the shebang from this file as it's
not a common approach to execute Rakefiles directly. It also updates the
approach to loading Rake tasks to be the modern one.

It then sets a conditional against the command to disable the Rake task
to prevent the Rakefile crashing in environments with a default task
defined. Finally it updates the ordering of the default to run lint
first as per other GOV.UK applications.