Remove ability to specify different button text
The functionality to specify button text on questions pages was added to
change the text to "Continue" for the "Find coronavirus support" flow.
Recently we've changed all Smart Answer flows to use the question button
text to be "Continue" instead of "Next step". This made this
functionality unused. Removing to simplfy the codebase and prevent
snowflake patterns from being introduced in the future.