Split student loan tests by year
For the tests that vary each year, this separates the tests into
context blocks for each year. This should make next year's uprating
easier, as the entire previous year's block can be deleted and the
current block copied to be updated for the following year, rather than
the current process of having to go through each test to work out what
needs changing for that year.